file: convert pmemfile_block->size to uint32_t
We'll never allocate blocks bigger than 2GB.

+++ src/libpmemfile/corelib/file.h
@@ -74,7 +74,7 @@ struct pmemfile_file {
 		unsigned block_id;
 
 		/* Offset from the beginning of current block. */
-		size_t block_offset;
+		uint32_t block_offset;
 
 		/* Above maps to below. */
